+++ src/libmain-c/.version
@@ -1,0 +1,1 @@
+../../.version
