+++ src/libmain-c/build-utils-meson
@@ -1,0 +1,1 @@
+../../build-utils-meson
