+++ src/libmain/plugin.hh
@@ -1,0 +1,12 @@
+#pragma once
+///@file
+
+namespace nix {
+
+/**
+ * This should be called after settings are initialized, but before
+ * anything else
+ */
+void initPlugins();
+
+}
